requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.6
packaging==24.0
requests==2.31.0
urllib3==2.2.1
svglib==1.6.0
reportlab>=4.1.0
pillow>=10.2.0
python-dateutil>=2.8.2
pyinstaller>=6.19.0
